WeeChat: Remove CV
CV, or Collegium V, was probably the first IRC server I idled on for a
long time.  Current network management no longer wishes to maintain
the IRC server and rather than riding it into the ground I parted
early.  Gone forever are the days of high adventure and engineering at
the edge of our capacity, but it was fun while it lasted!